refactor: refactor metadata and other helpers to support new features
