+++ apps/ui/src/types/general.ts
@@ -4,6 +4,12 @@ import { routing } from "@/lib/navigation"
 
 import { IntlKeysPath } from "./helpers"
 
+// Use type safe message keys with `next-intl`
+type Messages = typeof import("../../locales/en.json")
+
+// eslint-disable-next-line no-unused-vars
+export interface IntlMessages extends Messages {}
+
 export type AppLocale = (typeof routing.locales)[number]
 
 export interface AppError {
@@ -20,3 +26,9 @@ export interface AppLink {
   label?: string
   translateKey?: IntlKeysPath
 }
+
+export type NextMetadataTwitterCard =
+  | "summary"
+  | "summary_large_image"
+  | "player"
+  | "app"
